requirements.txt
pandas 
numpy 
matplotlib
scikit-learn
pyaml
mlflow
xgboost
catboost
lightgbm
sqlalchemy
imbalanced-learn
mysql-connector-python
requests
fastapi
uvicorn
dagshub
pytest
python-multipart
azure-storage-blob==12.19.1

# -e . # installs the current project in editable mode so local code changes reflect immediately without reinstalling.